Initialized Project Repositry
The main reason for creating this project is to learn how to scrap a
website and get the required data. As an additional extension, I will
also be learning to learn how to send email using python.

I will be using hackernews website to scrap. I will get all the post
with over 100 likes and send it to a email with a formatted way.